update to 0.7, refactor scattering, add tests

+++ src/plotting.jl
@@ -12,10 +12,10 @@ Given a classifier and a layeredTransform, it plots
 """
 function plotShattered(output::Array{Float64,3}, layers::layeredTransform; m=2, p=2, scale=:log, saveTo="", title="Shattering Transform Coefficients", aspect_ratio=1)
   average = [sqrt(norm(output[:,:,i], 2)) for i=1:size(output,3)]
-  xNames = Array{Int64}(layers.shears[m].shearletIdxs[1:end-1,:])
+  xNames = Array{Int}(layers.shears[m].shearletIdxs[1:end-1,:])
   xNames = ["$(xNames[i,:])  " for i=1:size(xNames,1)]
   n = size(layers.shears[m].shearletIdxs,1)-1
-  yNames = Array{Int64}(layers.shears[m-1].shearletIdxs[1:end-1,:])
+  yNames = Array{Int}(layers.shears[m-1].shearletIdxs[1:end-1,:])
   yNames = ["$(yNames[i,:])  " for i=1:size(yNames,1)]
   k = size(layers.shears[m-1].shearletIdxs,1)-1
   # pyplot()
@@ -43,11 +43,11 @@ end
 
 
 """
-    plotCoordinate(index::Vector{Array{Int64,1}}, m::Array{Int64}, layers::layeredTransform; fun=logabs, saveTo="")
+    plotCoordinate(index::Vector{Array{Int,1}}, m::Array{Int}, layers::layeredTransform; fun=logabs, saveTo="")
 
 given an index as listed in layers (starting with the second first), and a depth m, plot the corresponding shearlets, or if also given a scatteredND, plot the actual output for that coordinate. fun should be either abs, real, or imag
 """
-function plotCoordinate(index::Vector{Array{Int64,1}}, m::Array{Int64}, layers::layeredTransform; fun=logabs, saveTo="")
+function plotCoordinate(index::Vector{Array{Int,1}}, m::Array{Int}, layers::layeredTransform; fun=logabs, saveTo="")
   i=1
   for i = 1:size(layers.shears[m[1]].shearletIdxs,1)
     if layers.shears[m[1]].shearletIdxs[i,:]==index[1]
@@ -67,7 +67,7 @@ function plotCoordinate(index::Vector{Array{Int64,1}}, m::Array{Int64}, layers::
   return plt
 end
 
-function plotCoordinate(index::Vector{Array{Int64,1}}, m::Int64, layers::layeredTransform, sheared::S; fun=abs, title="", scale=:none, saveTo="") where S<:scattered
+function plotCoordinate(index::Vector{Array{Int,1}}, m::Int, layers::layeredTransform, sheared::S; fun=abs, title="", scale=:none, saveTo="") where S<:scattered
   # i is the index from the first to the last
   i=[1 for j=1:m]
   for j=1:m
@@ -90,5 +90,5 @@ end
 # TODO: A version of plotCoordinate that works for a list of indices
 # playing around with a way to automatically get square layouts
 # r=3
-# r1 = Int64(round(sqrt(r)))
-# Int64(ceil(r/r1))
+# r1 = Int(round(sqrt(r)))
+# Int(ceil(r/r1))
